knative/ChatBot/chat-split/requirements.txt
parliament-functions==0.1.0
Flask==2.2.3
minio==7.1.13
numpy==1.24.2
scipy==1.10.1
